refactor: move firewall implementations into platform-specific files
Replace runtime.GOOS dispatch with build-tagged files so each platform
has a proper FirewallState()/FirewallStealthState() implementation.
Eliminates stub functions for cross-platform compilation.
